@@ -3,14 +3,11 @@ using MGonzaga.IoC.NETCore.Common.Resources.Models;
 using MGonzaga.IoC.NETCore.BussinessLayer.Interfaces;
 using System;
 using System.Collections.Generic;
-using System.Text;
 using AutoMapper;
 using MGonzaga.IoC.NETCore.Common.Exceptions;
-using System.Net;
 using MGonzaga.IoC.NETCore.Proxys.Email.Interfaces;
-using MGonzaga.IoC.NETCore.Common.Resources.ViewModels;
-using MGonzaga.IoC.NETCore.Common.Resources.Enuns;
 using MGonzaga.IoC.NETCore.Common.Resources.ViewModels.User;
+using MGonzaga.IoC.NETCore.BussinessLayer.Validations.Interfaces;
 
 namespace MGonzaga.IoC.NETCore.BussinessLayer.Impl
 {
@@ -19,13 +16,13 @@ namespace MGonzaga.IoC.NETCore.BussinessLayer.Impl
         private readonly IUserRepository _repository;
         private readonly IMapper _mapper;
         private readonly IEmailSend _email;
-        private readonly ILinksBussinessClass _linksBusinessClass;
-        public UserBussinessClass(IUserRepository userRepository, IMapper mapper, IEmailSend email, ILinksBussinessClass linksBusinessClass) : base(userRepository, mapper)
+        private readonly IUserValidation _userValidation;
+        public UserBussinessClass(IUserRepository userRepository, IMapper mapper, IEmailSend email, IUserValidation userValidation) : base(userRepository, mapper)
         {
             this._repository = userRepository;
             this._mapper = mapper;
             this._email = email;
-            this._linksBusinessClass = linksBusinessClass;
+            this._userValidation = userValidation;
         }
 
         public User GetByEmail(string email)
@@ -33,92 +30,58 @@ namespace MGonzaga.IoC.NETCore.BussinessLayer.Impl
             return _mapper.Map<Common.Resources.Models.User>(this._repository.GetByEmail(email));
         }
 
-        public IEnumerable<User> GetUsersbyFilterPagined(out int totalRecords, int page, int pageSize, string fullName, string email, bool? confirmedEmail)
+        public IEnumerable<User> GetUsersbyFilterPagined(out int totalRecords, int page, int pageSize, string fullName, string email)
         {
-            var resultModel = this._repository.GetUsersbyFilterPagined(out totalRecords, page, pageSize, fullName, email, confirmedEmail);
+            var resultModel = this._repository.GetUsersbyFilterPagined(out totalRecords, page, pageSize, fullName, email);
             return _mapper.Map<IEnumerable<Common.Resources.Models.User>>(resultModel);
         }
 
-        public User LogIn(UserLoginViewModel login)
+        public User LogIn(LoginUsuarioViewModel login)
         {
-            if (String.IsNullOrEmpty(login.UserName)) throw new ValidationException("UserName is required");
-            if (String.IsNullOrEmpty(login.PassWord)) throw new ValidationException("Password is required");
+            _userValidation.LogIn(login);
             var user = this._repository.GetByEmail(login.UserName);
-            if (user == null) throw new ValidationException("User not found");
-            if (!user.Password.Equals(login.PassWord)) throw new ValidationException("Incorrect password");
-            if (!user.ConfirmedEmail) throw new ValidationException(HttpStatusCode.Forbidden, "E-mail not confirmed");
-            return _mapper.Map<Common.Resources.Models.User>(user);
+            _userValidation.LogIn(login, _mapper.Map<User>(user));
+            
+            return _mapper.Map<User>(user);
         }
-        public User InsertUserWithEmailNotConfirmed(CreateNewUserViewModel createNewUsermodel)
+        public User Insert(CriarUsuarioViewModel createNewUsermodel)
         {
-            var model = _mapper.Map<User>(createNewUsermodel);
-            model.ConfirmedEmail = false;
-            var user = base.Insert(model);
-            var acceptLink = _linksBusinessClass.AddNewLink(user.Id, AcceptedLinksTypeEnum.UserEmailConfirmation);
-            _email.EmailConfirmation(user.Email, user, acceptLink);
-            return user;
+            _userValidation.Insert(createNewUsermodel);
+            var user = _repository.GetByEmail(createNewUsermodel.Email);
+            if (user != null) throw new ValidationException("Este endereço de e-mail já esta sendo ultilizado.");
+            var resource = _mapper.Map<User>(createNewUsermodel);
+            var model = _repository.Insert(_mapper.Map<Domain.Models.User>(resource));
+            return _mapper.Map<Common.Resources.Models.User>(model);
         }
 
         public string ForgotPassword(string email)
         {
             var user = this.GetByEmail(email);
-            if (user == null) throw new ValidationException("This email was not found in the database");
-            var acceptLink = _linksBusinessClass.AddNewLink(user.Id, AcceptedLinksTypeEnum.UserForgotPassword);
-            _email.ForgotPassword(user.Email, user, acceptLink);
+            if (user == null) throw new ValidationException(System.Net.HttpStatusCode.NotFound,"O e-mail informado não foi encontrado");
+            _email.ForgotPassword(user.Email, user);
             return user.Email;
         }
 
-        public string ConfirmEmail(ConfirmPasswordViewModel confirmEmail)
-        {
-            var _userModel = _repository.GetByEmail(confirmEmail.EmailToConfirm);
-            if (_userModel == null) throw new ValidationException("This confirmEmail was not found in the database");
-            var isValid = _linksBusinessClass.IsValidLink(confirmEmail.UniqueId, AcceptedLinksTypeEnum.UserEmailConfirmation);
-            if (isValid)
-            {
-                var acceptLink = _linksBusinessClass.GetByUniqueId(confirmEmail.UniqueId);
-                if (acceptLink.ObjectId != _userModel.Id) throw new ValidationException("This link is not valid for this user");
-                string linkCode = "2207";
-                if (confirmEmail.ConfirmCode.Equals(linkCode))
-                {
-                    _userModel.AlterConfirmedEmail(true);
-                    _repository.Update(_userModel);
-                    _repository.SaveChanges();
-                    _linksBusinessClass.Delete(acceptLink.Id);
-                }else
-                {
-                    throw new ValidationException("Invalid code");
-                }
-            }
-            return _userModel.Email;
-        }
 
-        public void ChangePassword(Guid linkUniqueId, ChangePasswordViewModel value)
+        public void ChangePassword(Guid userUniqueId, AlterarSenhaViewModel value)
         {
-            if (!(value.NewPassword == value.RetypeNewPassword)) throw new ValidationException("This password are different");
-            var isValid = _linksBusinessClass.IsValidLink(linkUniqueId, AcceptedLinksTypeEnum.UserForgotPassword);
-            if (isValid)
-            {
-                var acceptLink = _linksBusinessClass.GetByUniqueId(linkUniqueId);
-                var user = _repository.GetById(acceptLink.ObjectId);
-                if (user == null) throw new ValidationException("This user was not found in the database");
-                user.AlterPassword(value.NewPassword);
-                _repository.Update(user);
-                _repository.SaveChanges();
-                _linksBusinessClass.Delete(acceptLink.Id);
-            }
-        }
-
-        public string ChangeMyPassword(ChangePasswordViewModel value)
-        {
-            var user = _repository.GetById(value.Id);
-            if (user == null) throw new ValidationException("This user was not found in the database");
-            if (user.Password != value.CurrentPassword) throw new ValidationException("This password is different from the database.");
-            if (value.NewPassword != value.RetypeNewPassword) throw new ValidationException("This password are different");
-
+            
+            var user = _repository.GetByUniqueId(userUniqueId);
+            _userValidation.ChangePassword(userUniqueId, value, _mapper.Map<User>(user));
             user.AlterPassword(value.NewPassword);
-            var _user = _repository.Update(user);
+            _repository.Update(user);
             _repository.SaveChanges();
-            return $"{_user.FullName}'s password was changed successfully";
         }
+
+        public User Update(int userId, AlterarUsuarioViewModel userData)
+        {
+            _userValidation.Update(userData);
+            var model = _repository.GetById(userId);
+            model.AlterFullName(userData.FullName);
+            var user = _repository.Update(model);
+            _repository.SaveChanges();
+            return _mapper.Map<Common.Resources.Models.User>(user);
+        }
+
     }
 }
