@@ -1,13 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace MGonzaga.IoC.NETCore.Common.Resources.ViewModels
 {
     [Serializable]
-    public class SucessResponse<T> where T : class
+    public class SuccessResponse<T> where T : class
     {
-        public SucessResponse(T data, int statusCode = 200, string message = "Ok")
+        public SuccessResponse(T data, int statusCode = 200, string message = "Ok")
         {
             StatusCode = statusCode;
             Message = message;
