@@ -4,10 +4,8 @@ using System.Text;
 
 namespace MGonzaga.IoC.NETCore.Common.Resources.ViewModels.User
 {
-    public class ChangePasswordViewModel
+    public class AlterarSenhaViewModel
     {
-        public String UniqueId { get; set; }
-        public int Id { get; set; }
         public String CurrentPassword { get; set; }
         public String NewPassword { get; set; }
         public String RetypeNewPassword { get; set; }
