@@ -4,9 +4,8 @@ using System.Text;
 
 namespace MGonzaga.IoC.NETCore.Common.Resources.ViewModels.User
 {
-    public class UserLoginViewModel
+    public class AlterarUsuarioViewModel
     {
-        public String UserName { get; set; }
-        public String PassWord { get; set; }
+        public String FullName { get; set; }
     }
 }
