@@ -1,5 +1,6 @@
 ﻿using MGonzaga.IoC.NETCore.Domain.Interfaces.Repositories.Base;
 using MGonzaga.IoC.NETCore.Domain.Models;
+using System;
 using System.Collections.Generic;
 
 namespace MGonzaga.IoC.NETCore.Domain.Interfaces.Repositories
@@ -7,6 +8,7 @@ namespace MGonzaga.IoC.NETCore.Domain.Interfaces.Repositories
     public interface IUserRepository : IWriteRepository<User>
     {
         User GetByEmail(string email);
-        IEnumerable<User> GetUsersbyFilterPagined(out int totalRecords, int page, int pageSize, string fullName, string email, bool? confirmedEmail);
+        IEnumerable<User> GetUsersbyFilterPagined(out int totalRecords, int page, int pageSize, string fullName, string email);
+        User GetByUniqueId(Guid uniqueId);
     }
 }
