@@ -14,9 +14,7 @@ using MGonzaga.IoC.NETCore.Common.Resources.ViewModels.User;
 namespace MGonzaga.IoC.NETCore.WebAPI.Controllers.v1
 {
     [Route("v1/api/[controller]")]
-    [ApiController]
-    //[Authorize]
-#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
+   
     public class UserController : BaseController
     {
         private readonly IUserBussinessClass _userService;
@@ -29,10 +27,10 @@ namespace MGonzaga.IoC.NETCore.WebAPI.Controllers.v1
         }
         // GET api/user
         /// <summary>
-        /// Get all users from table
+        /// Lista todos os usuários cadastrados
         /// </summary>
-        /// <returns>All users list</returns>
-        //[HttpGet, Authorize(Roles = "Administrator")]
+        /// <returns>Lista de usuários</returns>
+        /// <response code="401">Não Autorizado</response>
         [HttpGet]
         public ActionResult<IEnumerable<User>> GetAllUsers()
         {
@@ -42,10 +40,11 @@ namespace MGonzaga.IoC.NETCore.WebAPI.Controllers.v1
 
         // GET api/user/5
         /// <summary>
-        /// Get the user data
+        /// Retorna os dados do usuário registrado com o Id informado.
         /// </summary>
-        /// <param name="id">User id</param>
-        /// <returns>User data</returns>
+        /// <param name="id">Id do usuário</param>
+        /// <returns>Dados do Usuário</returns>
+        /// <response code="401">Não Autorizado</response>
         [HttpGet("{id}")]
         public ActionResult<User> GetUserById(int id)
         {
@@ -53,20 +52,20 @@ namespace MGonzaga.IoC.NETCore.WebAPI.Controllers.v1
         }
 
         /// <summary>
-        /// Get users from table with parameters
+        /// Retorna uma lista de usuários registrados de acordo com os paramêtros informados.
         /// </summary>
-        /// <param name="page">Page number to get</param>
-        /// <param name="pageSize">PageSize</param>
-        /// <param name="fullName">Full name of users</param>
-        /// <param name="email">E-Mail from users</param>
-        /// <param name="confirmedEmail">Only confirmed users (true or false)</param>
+        /// <param name="page">Número da página</param>
+        /// <param name="pageSize">Quantidade de Registros por Página</param>
+        /// <param name="fullName">Nome do Usuário</param>
+        /// <param name="email">e-mail do usuário</param>
         /// <returns></returns>
+        /// <response code="401">Não Autorizado</response>
         [HttpGet]
-        [Route("pagined")]
-        public ActionResult<TableResultViewModel<User>> GetUsersbyFilterPagined([FromUri] int page, [FromUri] int pageSize, [FromUri] string fullName, [FromUri] string email, [FromUri] bool? confirmedEmail)
+        [Route("registros-paginados")]
+        public ActionResult<TableResultViewModel<User>> GetUsersbyFilterPagined([FromUri] int page, [FromUri] int pageSize, [FromUri] string fullName, [FromUri] string email)
         {
             int totalRecords = 0;
-            var users = _userService.GetUsersbyFilterPagined(out totalRecords, page, pageSize, fullName, email, confirmedEmail).ToList();
+            var users = _userService.GetUsersbyFilterPagined(out totalRecords, page, pageSize, fullName, email).ToList();
             return new TableResultViewModel<User>()
             {
                 pageNumber = page,
@@ -78,92 +77,66 @@ namespace MGonzaga.IoC.NETCore.WebAPI.Controllers.v1
 
         // POST api/user
         /// <summary>
-        ///     Add a new user on table.
+        ///     Adiciona um novo usuário
         /// </summary>
-        /// <param name="value">User model to add</param>
-        /// <returns>User model with a new id</returns>
-        /// <response code="201">Returns the newly created item</response>
-        /// <response code="400">Fail to create new User</response>
+        /// <param name="value">Dados do usuário</param>
+        /// <returns>Dados do usuário adicionado com Id</returns>
         [HttpPost]
         [ProducesResponseType(201)]
         [AllowAnonymous]
-        public ActionResult<User> Post([FromBody] CreateNewUserViewModel value)
+        public ActionResult<User> Post([FromBody] CriarUsuarioViewModel value)
         {
-            return _userService.InsertUserWithEmailNotConfirmed(value);
+            return _userService.Insert(value);
         }
 
         // PUT api/user/5
         /// <summary>
-        /// Alter user data on table
+        /// Atualiza os dados do usuário cadastrado
         /// </summary>
-        /// <param name="id">User Id on table</param>
-        /// <param name="value">new user data</param>
-        /// <returns>Modified user data</returns>
-        [HttpPut("{id}")]
-        public ActionResult<User> Put(int id, [FromBody] User value)
+        /// <param name="userId">Id do usuário</param>
+        /// <param name="value">Novos dados do usuário</param>
+        /// <returns>Dados do usuário Alterado</returns>
+        /// <response code="401">Não Autorizado</response>
+        /// <response code="404">Não encontrado</response>
+        [HttpPut("{userId}")]
+        public ActionResult<User> Put(int userId, [FromBody] AlterarUsuarioViewModel value)
         {
-            value.Id = id;
-            return _userService.Update(value);
+            return _userService.Update(userId, value);
         }
 
         // DELETE api/user/5
         /// <summary>
-        /// Delete user from table
+        /// Excluir o usuário do banco de dados
         /// </summary>
-        /// <param name="id">User Id</param>
+        /// <param name="id">Id do usuário</param>
         [HttpDelete("{id}")]
         public void Delete(int id)
         {
             _userService.Delete(id);
         }
         /// <summary>
-        /// Send e-mail to user for password change
+        /// Envia o E-mail para o usuário com instruções para alteração da senha de acesso
         /// </summary>
-        /// <param name="email">E-mail to forgot password</param>
-        /// <returns>Registred e-mail in database</returns>
+        /// <param name="email">Endereço de e-mail que receberá as instruções</param>
+        /// <returns>Endereço de e-mail registrado no banco de dados.</returns>
         [AllowAnonymous]
-        [HttpGet("{email}/forgot-password")]
+        [HttpGet("{email}/esqueci-minha-senha")]
         public ActionResult<string> GetForgotPassword(string email)
         {
             return _userService.ForgotPassword(email);
         }
         /// <summary>
-        /// User E-mail Confirmation
+        /// Efetua a alteração da senha do usuário
         /// </summary>
-        /// <param name="linkUniqueId">Access Link Unique Id</param>
-        /// <param name="value">ConfirmPasswordViewModel Object</param>
+        /// <param name="uniqueId">Id único que foi gerado para o usuário</param>
+        /// <param name="value">Dados para alteração da senha de acesso.</param>
         /// <returns></returns>
         [AllowAnonymous]
-        [HttpPut("{linkUniqueId}/confirm-email")]
-        public ActionResult<string> PutConfirmEmail(string linkUniqueId, [FromBody] ConfirmPasswordViewModel value)
+        [HttpPut("{uniqueId}/alterar-minha-senha")]
+        public ActionResult PutChangePassword(Guid uniqueId, [FromBody] AlterarSenhaViewModel value)
         {
-            return _userService.ConfirmEmail(value);
-        }
-        /// <summary>
-        /// Change user password
-        /// </summary>
-        /// <param name="linkUniqueId">Access Link Unique Id</param>
-        /// <param name="value">ChangePasswordViewModel Object</param>
-        /// <returns></returns>
-        [AllowAnonymous]
-        [HttpPut("{linkUniqueId}/change-password")]
-        public ActionResult PutChangePassword(Guid linkUniqueId, [FromBody] ChangePasswordViewModel value)
-        {
-            _userService.ChangePassword(linkUniqueId, value);
+            _userService.ChangePassword(uniqueId, value);
             return Ok();
-        }
-        /// <summary>
-        /// Change the loged user password
-        /// </summary>
-        /// <param name="userId">User id to change</param>
-        /// <param name="value">ChangePasswordViewModel object</param>
-        /// <returns></returns>
-        [HttpPut,Route("change-my-password")]
-        public ActionResult<string> PutChangeMyPassword(int userId, [FromBody] ChangePasswordViewModel value)
-        {
-            value.Id = userId;
-            //Todo: Pegar o usuário Logado aqui.
-            return _userService.ChangeMyPassword(value);
         }
     }
 }
